==> dulee/3-2. LightGBM.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns


from sklearn.model_selection import StratifiedKFold
from lightgbm.sklearn import LGBMClassifier

import optuna

from sklearn.metrics import roc_auc_score

# 데이터 로드

import os

colab = os.path.isdir('./sample_data')
mount = os.path.isdir('/content/drive')
if colab:
    if not mount:
        from google.colab import drive
        drive.mount('/content/drive')
    base_path = '/content/drive/Othercomputers/내 컴퓨터/5_ML_Project/dulee/'
    data_path = '/content/drive/Othercomputers/내 컴퓨터/5_ML_Project/data/'
else:
    base_path = ''
    data_path = '../data/'

submission_df = pd.read_csv(data_path + 'sample_submission.csv', index_col='id')

train = pd.read_csv(base_path + 'train.csv', index_col='id')
test = pd.read_csv(base_path + 'test.csv', index_col='id')

##### 변수 설정

X = train.drop(columns=['defects'])
y = train['defects']
X_test = test

# 모델 학습

### Light GBM

def optimizer(trial):
    num_leaves = trial.suggest_categorical('num_leaves', [4, 8, 16, 32])
    learning_rate = trial.suggest_float('learning_rate', 0.01, 0.3)
    n_estimators = trial.suggest_int('n_estimators', 100, 300)

    # 매번 for loop 마다 객체를 생성할 필요 없음
    lgbm = LGBMClassifier(num_leaves=num_leaves,
                      learning_rate=learning_rate,
                      n_estimators=n_estimators,
                      random_state=61)

    folds = StratifiedKFold(shuffle=True, random_state=61) # n_splits=5 (default값)
    scores = []
    for train_idx, val_idx in folds.split(X, y):      # index로 분리: memory 절약
        lgbm.fit(X.iloc[train_idx], y.iloc[train_idx]) # fit할 때마다 새로운 lgbm 모델이 만들어짐.

        y_true = y.iloc[val_idx]
        y_proba = lgbm.predict_proba(X.iloc[val_idx])[:, 1]
        scores.append(roc_auc_score(y_true, y_proba))
    return np.mean(scores)

study = optuna.create_study(direction="maximize")
study.optimize(optimizer, n_trials=10)

print("Best roc_auc_score: %.4f" % study.best_value)
print("Best params: ", study.best_trial.params) # best score일 때의 하이퍼파라미터들

##### best model

lgbm_best = LGBMClassifier(**study.best_trial.params,
                           random_state=61)
lgbm_best.fit(X, y)

y_proba_lgbm = lgbm_best.predict_proba(X_test)[:, 1]
y_proba_lgbm

# 제출

submission_df['defects'] = y_proba_lgbm
submission_df.to_csv(base_path + 'submission.csv')
submission_df

### 확인

pd.read_csv(base_path + 'submission.csv', index_col='id')
