==> dulee/4.Ensemble_models.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import roc_auc_score
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier
from sklearn.ensemble import VotingClassifier
from lightgbm import LGBMClassifier

import pickle

# 데이터 로드

import os

colab = os.path.isdir('./sample_data')
mount = os.path.isdir('/content/drive')
if colab:
    if not mount:
        from google.colab import drive
        drive.mount('/content/drive')
    base_path = '/content/drive/MyDrive/Colab Notebooks/5_ML_Project/dulee/'
    data_path = '/content/drive/MyDrive/Colab Notebooks/5_ML_Project/data/'
else:
    base_path = ''
    data_path = '../data/'

submission_df = pd.read_csv(data_path+'sample_submission.csv', index_col='id')

train = pd.read_csv(base_path + 'train.csv', index_col='id')
test = pd.read_csv(base_path + 'test.csv', index_col='id')

##### 변수 설정

X = train.drop(columns=['defects'])
y = train['defects']
X_test = test

K = 15

rf = RandomForestClassifier(
    min_samples_leaf=141,
    max_depth=9,
    max_features=0.68758604791617,
    random_state=61,
)
# rf = pickle.load(open(base_path + "rf_best_1.pickle", "rb"))
rf

# et = ExtraTreesClassifier(
#     n_estimators=700,
#     min_samples_leaf=2,
#     max_depth=16,
#     max_features=0.793614074795712,
#     min_samples_split=48,
#     random_state=61,
# )
et = pickle.load(open(base_path + "et_best_1.pickle", "rb"))
et

# lgbm = LGBMClassifier(
#     num_leaves=4,
#     learning_rate=0.16467547044594108,
#     n_estimators=230,
#     random_state=61,
#     force_col_wise=True,
# )
lgbm = pickle.load(open(base_path + "lgbm_best_1.pickle", "rb"))
lgbm

ensemble = VotingClassifier(
    [('rf', rf), ('et',et), ('lgbm', lgbm)],
    voting='soft',
    weights = [0.3, 0.3, 0.4],
)

def oof_predict(model):
    folds = StratifiedKFold(n_splits=K, shuffle=True, random_state=61)
    scores = []
    predict_probas = []
    for train_idx, val_idx in folds.split(X, y):
        model.fit(X.iloc[train_idx], y.iloc[train_idx])
        y_proba = model.predict_proba(X.iloc[val_idx])[:, 1]
        y_true = y.iloc[val_idx]
        scores.append(roc_auc_score(y_true, y_proba))

        predict_probas.append(model.predict_proba(X_test)[:, 1])
    print(f"# AUC {np.mean(scores):.5f}")
    return np.mean(predict_probas, axis=0)

ensemble.fit(X, y)
y_proba = ensemble.predict_proba(X_test)[:, 1]
y_proba_oof = oof_predict(ensemble)

submission_df['defects'] = y_proba
submission_df.to_csv(base_path + 'submission_ensemble.csv')
submission_df['defects'] = y_proba_oof
submission_df.to_csv(base_path + 'submission_ensemble_oof.csv')
submission_df
